requirements.txt
requests
playwright
python-dotenv
fake-useragent
beautifulsoup4
